Results with tweaked database to avoid Si# isomers
The new resonance structure generator recognizes
:Si-Si:   is same as  Si#Si
but we don't have thermo for the latter, so it was guessing
something wrong and using it. Now we guess something also
wrong but too high, and thus use the :Si-Si: resonance
structure instead, which should have better thermo.

I think this may make the model more like it used to be.